pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "dominodatalab-data"
version = "0.5.1dev1"
description = "Domino Data API for interacting with Domino Data features"
readme = "README.md"
authors = [
    "Gabriel Haim <gabriel.haim@dominodatalab.com>",
    "Aaron Read <aaron.read@dominodatalab.com>",
]
license = "Apache Software License 2.0"
repository = "https://github.com/dominodatalab/domino-data"
homepage = "https://github.com/dominodatalab/domino-data"

# Generated client need to be bundled
packages = [
    { include = "domino_data" },
    { include = "datasource_api_client" },
    { include = "training_set_api_client" },
    { include = "feature_store_api_client" },
]

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  #! Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  #! Update me
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
]

[tool.poetry.group.main.dependencies]
python = "^3.8"
pandas = ">=1.3.0"
httpx = "^0.23.0"
attrs = ">=20.1.0,<22.0.0"
python-dateutil = "^2.8.0"
pyarrow = "^8.0.0"
loguru = "^0.5.3"
backoff = "^1.11.1"
bson = "^0.5.10"

[tool.poetry.group.dev.dependencies]
Sphinx = "^4.1.2"
bandit = "^1.7.0"
black = {version = "^22.3.0", allow-prereleases = true}
darglint = "^1.8.0"
flake8 = "^4.0.1"
grpcio = "^1.41.0"
isort = {extras = ["colors"], version = "^5.9.3"}
mypy = "^0.910"
mypy-extensions = "^0.4.3"
openapi-python-client = "^0.11.1"
pdbpp = "^0.10.3"
pre-commit = "^2.13.0"
pydocstyle = "^6.1.1"
pylint = "^2.9.6"
pytest = "^6.2.4"
pytest-cov = "^3.0.0"
pytest-recording = "^0.12.0"
pyupgrade = "^2.23.1"
respx = "^0.19.1"
safety = "^2.3.1"
sphinx-autodoc-typehints = "^1.12.0"
sphinx-rtd-theme = "^0.5.2"
types-PyYAML = "^6.0.4"

[tool.poetry.group.featurestore]
optional = true

[tool.poetry.group.featurestore.dependencies]
feast = "^0.26.0"
GitPython = "^3.1.29"

[tool.black]
# https://github.com/psf/black
target-version = ["py38"]
line-length = 100
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 38
line_length = 100

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
extend_skip = ["services"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
# mypy configurations: https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
exclude = 'datasource_api_client/|domino_data/training_sets/'
python_version = 3.8
pretty = true
show_traceback = true
color_output = true

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = true
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.coverage.run]
omit = [
    "domino_data/_feature_store/run_featurestore_sync.py"
]
